feat(emacs): make emacs open file to a line like a normal editor
When opening files in emacs from the terminal using the default command `+LINE`
is not really useful an no programs output files and lines in that format. All
programmes output a file like `path/to/file.ext:2` this is line 2 in the file
`path/to/file.ext` if you were to open this with the emacs command it would be
`emacs +2 path/to/file.ext`. With this commit you can use `emacs
path/to/file.ext:2` and it will open the file on line 2 like you would expect.

As a extra I have overridden the tmux open command to add the `--no-wait` to the
editor command so that we can continue to use the terminal window as the file is
open in emacs and we don't need to close the file to use the terminal.