fix(git): format .gitconfig and move to Emacs for the editor
For some reason the .gitconfig was indented with tabs this has now been
correctly formatted. Also move over to the Emacs client for the git editor. This
has been working well for now. My in the future look at forcing the editor to
open in the terminal as sometimes the focusing in X is a bit off