refactor(shell): go back full terminal
After a long while in emacs gui trying the terminal inside emacs and using the
external terminal with tmux. I have now decided to back full terminal using
emacs in the terminal in a tmux pain, sort of the same way I was using vim
before I switched to emacs.