refactor(shell): refactor ggpushmr git alias
ggpushmr has now been renamed to ggpushpr. The only place it is called a "Merge
Request" is in Gitlab. All the community has now decided to go with Github's
"Pull Request".

This also makes the branch param optional. If you don't pass a branch name it
will make the pull request against the repositories default branch.