feat(shell): override "gc" alias to open up the ".git/COMMIT_EDITMSG"
When using tools like commitlint to lint your commits there is always that time
you write an invalid commit. In this situation you are dumped back into the
terminal with an error message and your invalid commit.

This function will now load the invalid commit message back in to the editor so
you can fix the errors and commit successfully.