refactor(shell): remove emacs zsh plugin and create my own
This dose the same thing as the built in oh-my-zsh plugin but this one can
update the paths to emacs that it will be used.

This will allow you to use a custom compiled version of emacs for testing and /
or pinning the version you want with the features you need.