feat: ask to remove branch in git-merge-patch
When trying to merge-patch it will fail if the branch you are trying to merge is
already on the local machine.

This will ask if you want to remove it. The branch will be removed locally and
then pulled from the remote to ensure you are not merging any un-pushed local
changes.