chore(pp): move to neovim unstable
This is to get some features to use some plugins like the built in LSP.
On the hole its going well. Writing code is way more responsive with
the built in LSP. Project navigation is still not there, Ivy is a really
nice emacs package that I am still trying to find a replacement.