fix(vim): lint files when we open buffers
Currently they are only getting saved when you save buffers. I was
missing errors when jumping around because the files do not get saved.

This will now show me the errors when the buffer opens not only when the
buffer is saved and I get to see linting errors when jumping around a
codebase.