feat(vim): improve markdown editing with long wrapping lines
In markdown I have long lines wrapping as we are often editing docs that
don't stick to 80 columns. This makes it hard to navigate with j and k
imposable to use becase is skips over the wrapped lines.

This now using gj and gk to navigate wrapped lines so j and k will act
naturaly and move down one "line" even on the wrapping.