feat(vim): introduce key mappings for folding
This enhancement includes key mappings designed to improve code folding
in Vim. One of these mappings involves a double leader, which
efficiently toggles the current fold. Additionally, another mapping
utilizes the square bracket convention for seamless navigation between
folds, employing the mnemonic "fold."

Consideration was given to the potential conflict between the `f` key
and "function." After thoughtful deliberation, the decision was made to
use LSP for navigating around functions.

The implementation is currently integrated with prr to narrow the focus
of differentials to a single hunk. This refinement significantly
enhances the manageability of reviewing large pull requests.