feat(vim): added neoformat plugin to configuration
I have added the neoformat plugin to the list of plugins in my Neovim
configuration file. This plugin provides a powerful formatting tool that
can be used to format code and files in various programming languages.
It also supports multiple formatting options, making it easy to choose
the right settings for your specific needs.

I have been using this plugin for a while now, I must have missed it
when migrating to configz