chore(vim): remove async from the default test
We are now adding lints test that disable async tests that do not use
await in them. I am by far the biggest offender for this and this is one
of the main reasons. Now any test I write will be synchronous by default
until I need it to be async.