feat(sapling): improve log
Change the included commits to be the full tree, no matter where you are
on the stack. There is the graphnode to indicate where you are with a
`@`.

Change the color of public commits so you can easily see what are your
commits on the stack. This makes moving around and rebasing easter when
you can see when your last public commit it.

Remove the hard 40 commit limit. Now we are getting the last 40 public
commits plus the commits on your stack