fix(vim): add `--no-require-git` to the ferret rg command
Now that I have mostly moved over to sapling, rg dose not ignore git
files by default. When searching in vim it would search in the
node_modules by default. As you can imagine this took for ever.

Now we are passing the `--no-require-git` flag so it will ignore any
file / dirs that are in any .gitignore by default.