fix(core): issues with the install script
For the zypper install script, there was a spelling mistake in
`libopenssl-devel`, this is required for compiling configz.

On WSL git is not installed by default, this now installs it so we can
download the stuff.

In the conditions to see if packages are installed, there are no quotes
around the commands. This is then giving false positives and not
installing configz or the rust toolchain. This has now been sorted so
its all good.