feat(vim): add baleia.nvim plugin to color buffers
This plugin will allow you to color a buffer with ansi characters. This
is good for when you want to dump the output of a command into a buffer,
you can do it with nice colors now too.

Sapling.nvim uses this for the log buffers. That is not installed with
our dotfiles, I will someday have a sapling provider for installing and
keeping plugins updated, for now I will keep it updated manually for
now.