fix(tmux): set the actual name to the tmux plugin
This swap over should have really been tested more before pushing into
mainline. This is hopefully the last one we need to fix.

When I was playing around with the plugin I used a `git clone` this have
it the name of the repo "tmux" that was not very helpful when looking at
the plugins. I then renamed it when adding it to the dotfiles install.
However, I never changed the name in the config to load the plugin from
a different place.