fix(git): navigation in diffs
When using delta diff, you could not navigate the hunks of the diff. It
would only jump between file changes. This was due to the use of the raw
hunk header. It must not be included in the default `navigate-regex`.

I have updated the regex so it includes the default hunk patter so now
when you have in a diff and press `n` it will navigate the hunks as well
as the file changes.