chore: add github git auth helper
Adds the auth helper to the config. This is used mainly for sapling so
we can push and pull to git repos. I need to run this after each time I
configure the system because this file is a template not a symlink

This was originally added with the `gh` cli tool running the following
command:

```bash
gh auth setup-git
```