feat(wezterm): add <ctrl-bb> binding to go to the previous tab
When developing I often want to quickly swap between vim and a terminal. The
vim built in terminal does not really work for be. I prefer to use a proper
terminal.

Now you can quickly swap by pressing <LEADER-b> this is the same kind of
workflow I have in vim by using the <LEADER-TAB>