feat(vim): add the deno lsp to the config
This is a bit tricky due to the filetype of deno being typescript. The
typescript server automatically starts on this filetype.

For now this looks in the root of the project to see if there is a deno.json or
package.json. Based on this, it will pick what one to use.