feat(emacs): add / key binding to search the roam notes
Usually this mapping would be to live grep in the current project. With emacs I
only have one project and thats my notes. The sole purpose of emacs for me is
to organize my self with org

Now this binding search for content in the org notes directory treating that as
the project.