feat(emacs): show only long term scheduled tasks not all of them
When viewing the org agenda I have a section to view all the scheduled
tasks. That is not that helpful as most of them are shown in the agenda
view on the calendar.

This now limits the scheduled tasks to be only tasks that are not on the
calendar view for this week, to remove duplicate task from being shown.