fix(nu): support windows nicely
The main thing we are doing is setting the `HOME` and `USER` env variables so
we can use comon variables over linux and windows.

There is also a windows wezterm issue `shell_integration.osc133` feature needs
to be disabled. This will stop the prompt adding a new line on every char typed
into the prompt.