fix(terminal): wezterm nvim switch pane integration
When working on windows the top most process name is the current foreground
process name. For example if you are runing something like a terminal in nvim
the terminal process is the foreground like `nu`

This now gets the full process stack and if any of the processes in the stack
are a editor then it will send the binding to the editor to change the pane.

This now will give you windows wezterm integration for the tmux navigator
custom plugins.